Moved split_top_level_commas function for jump box input parsing and update multi_entry_jump_search to use it as default separator
