+++ lmfdb/utils/search_wrapper.py
@@ -1,4 +1,3 @@
-import re
 import time
 from random import randrange
 from flask import render_template, jsonify, redirect, request, url_for
@@ -40,8 +39,38 @@ def use_split_ors(info, query, split_ors, offset, table):
     )
 
 
+def split_top_level_commas(text):
+    """
+    A function which takes an input string and returns a list of strings, splitting on commas that are not inside parentheses/brackets/braces.
+    Used as the default separator function when parsing jump box input for multiple entries.
+    """
+
+    entries = []
+    chunk = []
+    depth = 0
+    for ch in text:
+        if ch in "([{":
+            depth += 1
+            chunk.append(ch)
+        elif ch in ")]}":
+            depth = max(depth - 1, 0)
+            chunk.append(ch)
+        elif ch == "," and depth == 0:
+            entry = "".join(chunk).strip()
+            if entry:
+                entries.append(entry)
+            chunk = []
+        else:
+            chunk.append(ch)
+
+    entry = "".join(chunk).strip()
+    if entry:
+        entries.append(entry)
+    return entries
+
+
 def multi_entry_jump_search(info, parse_entry, label_exists, index_endpoint, input_key="jump", labels_key="labels", 
-                            sep=lambda x: re.split(",", x), object_name="records", timeout="30"):
+                            sep=split_top_level_commas, object_name="records", time_limit=30):
     """
     Generic handler for jump boxes that supports comma-separated input of various entries (labels/names/polynomials/equations etc.).
 
@@ -56,9 +85,9 @@ def multi_entry_jump_search(info, parse_entry, label_exists, index_endpoint, inp
     - ``index_endpoint`` -- the input to "url_for" which returns the index homepage for this section
     - ``input_key`` -- the dictionary key for the jump search box (default: "jump")
     - ``labels_key`` -- the dictionary key for the labels search query (default: "labels")
-    - ``sep`` -- A function used to seperate out the jump box input into seperate entries (default: lambda x: re.split(",", x))
+    - ``sep`` -- A function used to separate out jump box input into separate entries
     - ``object_name`` -- The name of the objects in the database (e.g. "fields", "elliptic curves"). Used when flashing info or error messages.
-    - ``time_limit`` -- a time limit (in seconds) for the maximaum amount of time this query should take
+    - ``time_limit`` -- a time limit (in seconds) for the maximum amount of time this query should take
     """
 
     jump_input = info.get(input_key, "")
@@ -66,16 +95,16 @@ def multi_entry_jump_search(info, parse_entry, label_exists, index_endpoint, inp
     if len(entries) <= 1:
         return None
 
-    # For each entry given in the comma-seperated jump box input, we attempt to parse the entry using parse_entry (whilst skipping over duplicates)
-    # If the user inputs a large of entries, this may take a long time (e.g. for number fields, this might require calling polredabs on every entry)
+    # For each entry given in the comma-separated jump box input, we attempt to parse the entry using parse_entry (while skipping duplicates)
+    # If the user inputs a large number of entries, this may take a long time (e.g. for number fields, this might require calling Pari's polredabs on every entry)
     # We start a timer and stop parsing entries if we've hit the specified time_limit (default: 30 seconds)
 
     labels, seen = [], set()
     not_parsed, not_found = 0, 0
-    start_timer = time.time()
+    start_timer = time.monotonic()
     for i in range(len(entries)):
         # Check if exceeded time limit
-        if time.time() - start_timer > timeout:
+        if time.monotonic() - start_timer > time_limit:
             flash_error("Query timed out after parsing the first %s entries in the input.", i)
             return redirect(url_for(index_endpoint))
 
@@ -92,7 +121,7 @@ def multi_entry_jump_search(info, parse_entry, label_exists, index_endpoint, inp
             labels.append(label)
             seen.add(label)
 
-    # Flash error if no entries succesfully parsed
+    # Flash error if no entries successfully parsed
     if not labels:
         flash_error("None of the %s entries matched %s in the database.", len(entries), object_name)
         return redirect(url_for(index_endpoint))
